datetime to UTC
