+++ Tracker_API/auth/routes.py
@@ -1,9 +1,4 @@
 from datetime import datetime, timedelta
-# import pytz
-  
-# # it will get the time zone 
-# # of the specified location
-# IST = pytz.timezone('Asia/Kolkata')
 
 from . import auth
 from .. import db
@@ -31,12 +26,12 @@ def login():
     if check_password_hash(user.Password, form_data["Password"]):
         token = jwt.encode({
                             'Username' : user.Username, 
-                            'exp' : datetime.now() + timedelta(seconds=30)
+                            'exp' : datetime.utcnow() + timedelta(minutes=30)
                             },
                         current_app.config['SECRET_KEY'],
                         algorithm="HS256"
                         )
-        return jsonify({'token' : token, 'expires_at' : datetime.now() + timedelta(seconds=30)}), 200
+        return jsonify({'token' : token, 'expires_at' : datetime.utcnow() + timedelta(minutes=30)}), 200
 
     return make_response('Could not verify', 401, {'WWW-Authenticate' : 'Basic realm="Login required!"'})
