+++ Tracker_API/models.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from multiprocessing import Event
 from flask_login import UserMixin
 from . import db
 
@@ -9,13 +8,13 @@ class tbl_events(db.Model):
     EventName = db.Column(db.String(100), unique=True, nullable=False)
     EventDescription = db.Column(db.String(100), nullable=True)
     NumberOfMembers = db.Column(db.Integer, nullable=False)
-    EventTime = db.Column(db.DateTime, nullable=False)
+    EventTime = db.Column(db.DateTime(timezone=True), nullable=False)
     txns = db.relationship('tbl_tlist', backref='txn_event', lazy=True)
     event_users = db.relationship('tbl_users', secondary='tbl_eventusers', back_populates='user_events', lazy=True)
     event_txnShare = db.relationship('tbl_txnshare', backref='txnShare_event', lazy=True)
 
     def __init__(self, EventName, EventDescription=None, NumberOfMembers=0,
-                     EventTime=datetime.now(), ):
+                     EventTime=datetime.utcnow(), ):
         self.EventName = EventName
         self.EventDescription = EventDescription
         self.NumberOfMembers = NumberOfMembers
@@ -30,11 +29,11 @@ class tbl_tlist(db.Model):
     createdByUserID = db.Column(db.Integer, db.ForeignKey('tbl_users.id'), nullable=False) #txn_createdUser
     Amount = db.Column(db.Float, nullable=False)
     TxnDescription = db.Column(db.String(100), nullable=True)
-    TxnTime = db.Column(db.DateTime, nullable=False)
+    TxnTime = db.Column(db.DateTime(timezone=True), nullable=False)
     shared_users = db.relationship('tbl_users', secondary='tbl_txnshare', back_populates='user_txnShares', lazy=True)
 
     def __init__(self, EventID, paidByUserID, createdByUserID, Amount=0,
-                    TxnDescription=None, TxnTime=datetime.now(), ):
+                    TxnDescription=None, TxnTime=datetime.utcnow(), ):
         self.EventID = EventID
         self.paidByUserID = paidByUserID
         self.Amount = Amount
@@ -66,9 +65,9 @@ class tbl_eventusers(db.Model):
     __tablename___ = 'tbl_eventusers'
     EventID = db.Column(db.Integer, db.ForeignKey('tbl_events.EventID'), primary_key=True, nullable=False) #event
     UserID = db.Column(db.Integer, db.ForeignKey('tbl_users.id'), primary_key=True, nullable=False) #user
-    JoinTime = db.Column(db.DateTime, nullable=False)
+    JoinTime = db.Column(db.DateTime(timezone=True), nullable=False)
 
-    def __init__(self, EventID, UserID, JoinTime=datetime.now()):
+    def __init__(self, EventID, UserID, JoinTime=datetime.utcnow()):
         self.EventID = EventID
         self.UserID = UserID
         self.JoinTime = JoinTime
